+++ src/actions/index.ts
@@ -6,6 +6,12 @@ import { getSchema, getBaseIdFromId } from '../utils';
 import { DeleteWithi18nAction } from './DeleteWithi18nAction';
 import { DuplicateWithi18nAction } from './DuplicateWithi18nAction';
 
+export {
+  PublishWithi18nAction,
+  DeleteWithi18nAction,
+  DuplicateWithi18nAction,
+}
+
 export default (props: IResolverProps) => {
   const schema: Ti18nSchema = getSchema(props.type);
   const isI18n = schema && schema.i18n;