+++ src/types/TSchema.ts
@@ -1,5 +1,6 @@
 export type TSchema<T = any> = T & {
-    name: string;
-    title: string;
-    fields: any[];
+  name: string;
+  title: string;
+  icon?: any;
+  fields: any[];
 };